chore: move the apim auth logic to a more sensible place
